ai-models/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
joblib==1.3.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
pillow==10.0.0

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Data Processing
opencv-python==4.8.0.76
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.1.1

# Deployment
gunicorn==21.2.0
waitress==2.1.2
